Bugfix and change of methodology.
