+++ aws_sat_api/search.py
@@ -216,14 +216,3 @@ def sentinel2(utm: Union[str, int], lat: str, grid: str,
         results = executor.map(_info_worker, version_dirs)
 
     return results
-
-
-if __name__ == '__main__':
-    import timeit
-
-    def timing_func():
-        start_date = datetime(2017, 1, 1)
-        end_date = datetime(2017, 5, 15)
-        sentinel2(22, "K", "HV", start_date=start_date, end_date=end_date)
-    t = timeit.timeit('timing_func()', number=10, setup="from __main__ import timing_func")
-    print(t)
